User can only update & delete animal when them is the owner

+++ src/Feature/Animal/Handlers/DeleteAnimalHandler.cs
@@ -33,6 +33,13 @@ namespace Animal.Handlers
                 return DeleteAnimalResult.RequestEmptyResult();
             }
 
+            var userCanUpdateAnimal = await _animalService.UserOwnsAnimalAsync(request.Id, request.UserId);
+
+            if (!userCanUpdateAnimal)
+            {
+                return DeleteAnimalResult.AccessDeniedResult();
+            }
+
             var isDeleted = await _animalService.DeleteAnimalAsync(request.Id);
 
             return !isDeleted ? DeleteAnimalResult.BadRequestResult(request.Id) : DeleteAnimalResult.SuccessfulResult(request.Id);
